requirements.txt
Flask==0.10.1
flask-marshmallow==0.6.2
Flask-Migrate==1.8.0
Flask-RESTful==0.3.5
Flask-SQLAlchemy==2.1
marshmallow==2.6.0
marshmallow-jsonapi==0.5.0
marshmallow-sqlalchemy==0.8.1
PyYAML==3.11
